+++ pkg/plugin/bodylmt/setup.go
@@ -16,9 +16,9 @@ func init() {
 	})
 }
 
-func setupBodyLimit(route *proxy.Route, p plugin.Params) error {
+func setupBodyLimit(route *proxy.Route, rawConfig plugin.Config) error {
 	var config Config
-	err := plugin.Decode(p.Config, &config)
+	err := plugin.Decode(rawConfig, &config)
 	if err != nil {
 		return err
 	}