+++ pkg/plugin/rate/setup_test.go
@@ -3,12 +3,8 @@ package rate
 import (
 	"testing"
 
-	"github.com/garyburd/redigo/redis"
 	"github.com/hellofresh/janus/pkg/plugin"
 	"github.com/hellofresh/janus/pkg/proxy"
-	"github.com/hellofresh/janus/pkg/store"
-	stats "github.com/hellofresh/stats-go"
-	"github.com/rafaeljusto/redigomock"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -42,13 +38,8 @@ func TestRateLimitPluginLocalPolicy(t *testing.T) {
 		"policy": "local",
 	}
 
-	statsClient, _ := stats.NewClient("memory://", "")
 	route := proxy.NewRoute(&proxy.Definition{})
-	err := setupRateLimit(route, plugin.Params{
-		Config:      rawConfig,
-		Storage:     store.NewInMemoryStore(),
-		StatsClient: statsClient,
-	})
+	err := setupRateLimit(route, rawConfig)
 
 	assert.NoError(t, err)
 	assert.Len(t, route.Inbound, 2)
@@ -60,36 +51,8 @@ func TestRateLimitPluginRedisPolicyWithInvalidStorage(t *testing.T) {
 		"policy": "redis",
 	}
 
-	statsClient, _ := stats.NewClient("memory://", "")
 	route := proxy.NewRoute(&proxy.Definition{})
-	err := setupRateLimit(route, plugin.Params{
-		Config:      rawConfig,
-		Storage:     store.NewInMemoryStore(),
-		StatsClient: statsClient,
-	})
-
-	assert.Error(t, err)
-}
-
-func TestRateLimitPluginRedisPolicy(t *testing.T) {
-	rawConfig := map[string]interface{}{
-		"limit":  "10-S",
-		"policy": "redis",
-	}
-
-	pool := redis.NewPool(func() (redis.Conn, error) {
-		return redigomock.NewConn(), nil
-	}, 0)
-	storage, err := store.NewRedisStore(pool, "")
-	assert.NoError(t, err)
-
-	statsClient, _ := stats.NewClient("memory://", "")
-	route := proxy.NewRoute(&proxy.Definition{})
-	err = setupRateLimit(route, plugin.Params{
-		Config:      rawConfig,
-		Storage:     storage,
-		StatsClient: statsClient,
-	})
+	err := setupRateLimit(route, rawConfig)
 
 	assert.Error(t, err)
 }
@@ -100,13 +63,8 @@ func TestRateLimitPluginInvalidPolicy(t *testing.T) {
 		"policy": "wrong",
 	}
 
-	statsClient, _ := stats.NewClient("memory://", "")
 	route := proxy.NewRoute(&proxy.Definition{})
-	err := setupRateLimit(route, plugin.Params{
-		Config:      rawConfig,
-		Storage:     store.NewInMemoryStore(),
-		StatsClient: statsClient,
-	})
+	err := setupRateLimit(route, rawConfig)
 
 	assert.Error(t, err)
 }