+++ pkg/plugin/responsetransformer/setup.go
@@ -11,9 +11,9 @@ func init() {
 	})
 }
 
-func setupResponseTransformer(route *proxy.Route, p plugin.Params) error {
+func setupResponseTransformer(route *proxy.Route, rawConfig plugin.Config) error {
 	var config Config
-	err := plugin.Decode(p.Config, &config)
+	err := plugin.Decode(rawConfig, &config)
 	if err != nil {
 		return err
 	}