+++ pkg/web/provider.go
@@ -8,45 +8,46 @@ import (
 	"github.com/hellofresh/janus/pkg/api"
 	"github.com/hellofresh/janus/pkg/checker"
 	"github.com/hellofresh/janus/pkg/config"
-	"github.com/hellofresh/janus/pkg/errors"
+	httpErrors "github.com/hellofresh/janus/pkg/errors"
 	"github.com/hellofresh/janus/pkg/jwt"
 	"github.com/hellofresh/janus/pkg/middleware"
 	"github.com/hellofresh/janus/pkg/notifier"
-	"github.com/hellofresh/janus/pkg/oauth"
+	"github.com/hellofresh/janus/pkg/plugin"
 	"github.com/hellofresh/janus/pkg/router"
+	"github.com/pkg/errors"
 	"github.com/rs/cors"
 	log "github.com/sirupsen/logrus"
 )
 
-// Provider is a provider.Provider implementation that provides the REST API.
-type Provider struct {
-	Port     int  `description:"Web administration port"`
-	ReadOnly bool `description:"Enable read only API"`
-	Cred     config.Credentials
-	Notifier notifier.Notifier
-	TLS      config.TLS
-	APIRepo  api.Repository
-	AuthRepo oauth.Repository
+func init() {
+	plugin.RegisterEventHook(plugin.StartupEvent, onStartup)
 }
 
-// Provide executes the provider functionality
-// This is normally the entry point of the
-// provider.
-func (p *Provider) Provide(version string) error {
-	log.Info("Janus Admin API starting...")
+func onStartup(event interface{}) error {
+	e, ok := event.(plugin.OnStartup)
+	if !ok {
+		return errors.New("Could not convert event to startup type")
+	}
 
-	router.DefaultOptions.NotFoundHandler = errors.NotFound
+	config := e.Config.Web
+
+	repo, err := api.BuildRepository(e.Config.Database.DSN, e.MongoSession)
+	if err != nil {
+		return err
+	}
+
+	log.Info("Janus Admin API starting...")
+	router.DefaultOptions.NotFoundHandler = httpErrors.NotFound
 	r := router.NewChiRouterWithOptions(router.DefaultOptions)
 
 	// create authentication for Janus
-	guard := jwt.NewGuard(p.Cred)
-	authMiddleware := jwt.NewMiddleware(guard)
+	guard := jwt.NewGuard(config.Credentials)
 	r.Use(
 		chimiddleware.StripSlashes,
 		chimiddleware.DefaultCompress,
 		middleware.NewLogger().Handler,
-		middleware.NewRecovery(errors.RecoveryHandler),
-		middleware.NewOpenTracing(p.TLS.IsHTTPS()).Handler,
+		middleware.NewRecovery(httpErrors.RecoveryHandler),
+		middleware.NewOpenTracing(config.TLS.IsHTTPS()).Handler,
 		cors.New(cors.Options{
 			AllowedOrigins:   []string{"*"},
 			AllowedHeaders:   []string{"*"},
@@ -56,10 +57,10 @@ func (p *Provider) Provide(version string) error {
 	)
 
 	// create endpoints
-	r.GET("/", Home(version))
+	r.GET("/", Home())
 	// health checks
-	r.GET("/status", checker.NewOverviewHandler(p.APIRepo))
-	r.GET("/status/{name}", checker.NewStatusHandler(p.APIRepo))
+	r.GET("/status", checker.NewOverviewHandler(repo))
+	r.GET("/status/{name}", checker.NewStatusHandler(repo))
 
 	handlers := jwt.Handler{Guard: guard}
 	r.POST("/login", handlers.Login())
@@ -68,31 +69,31 @@ func (p *Provider) Provide(version string) error {
 		authGroup.GET("/refresh_token", handlers.Refresh())
 	}
 
-	p.loadAPIEndpoints(r, authMiddleware.Handler)
-	p.loadOAuthEndpoints(r, authMiddleware.Handler)
+	loadAPIEndpoints(r, repo, e.Notifier, guard)
+	plugin.EmitEvent(plugin.AdminAPIStartupEvent, plugin.OnAdminAPIStartup{Router: r})
 
 	go func() {
-		p.listenAndServe(r)
+		listenAndServe(config, r)
 	}()
 
 	return nil
 }
 
-func (p *Provider) listenAndServe(handler http.Handler) error {
-	address := fmt.Sprintf(":%v", p.Port)
+func listenAndServe(config config.Web, handler http.Handler) error {
+	address := fmt.Sprintf(":%v", config.Port)
 
 	log.Info("Janus Admin API started")
-	if p.TLS.IsHTTPS() {
-		addressTLS := fmt.Sprintf(":%v", p.TLS.Port)
-		if p.TLS.Redirect {
+	if config.TLS.IsHTTPS() {
+		addressTLS := fmt.Sprintf(":%v", config.TLS.Port)
+		if config.TLS.Redirect {
 			go func() {
 				log.WithField("address", address).Info("Listening HTTP redirects to HTTPS")
-				log.Fatal(http.ListenAndServe(address, RedirectHTTPS(p.TLS.Port)))
+				log.Fatal(http.ListenAndServe(address, RedirectHTTPS(config.TLS.Port)))
 			}()
 		}
 
 		log.WithField("address", addressTLS).Info("Listening HTTPS")
-		return http.ListenAndServeTLS(addressTLS, p.TLS.CertFile, p.TLS.KeyFile, handler)
+		return http.ListenAndServeTLS(addressTLS, config.TLS.CertFile, config.TLS.KeyFile, handler)
 	}
 
 	log.WithField("address", address).Info("Certificate and certificate key were not found, defaulting to HTTP")
@@ -100,40 +101,18 @@ func (p *Provider) listenAndServe(handler http.Handler) error {
 }
 
 //loadAPIEndpoints register api endpoints
-func (p *Provider) loadAPIEndpoints(router router.Router, handlers ...router.Constructor) {
+func loadAPIEndpoints(router router.Router, repo api.Repository, ntf notifier.Notifier, guard jwt.Guard) {
 	log.Debug("Loading API Endpoints")
 
 	// Apis endpoints
-	handler := api.NewController(p.APIRepo, p.Notifier)
+	handler := api.NewController(repo, ntf)
 	group := router.Group("/apis")
-	group.Use(handlers...)
+	group.Use(jwt.NewMiddleware(guard).Handler)
 	{
 		group.GET("/", handler.Get())
 		group.GET("/{name}", handler.GetBy())
-
-		if false == p.ReadOnly {
-			group.POST("/", handler.Post())
-			group.PUT("/{name}", handler.PutBy())
-			group.DELETE("/{name}", handler.DeleteBy())
-		}
-	}
-}
-
-//loadOAuthEndpoints register api endpoints
-func (p *Provider) loadOAuthEndpoints(router router.Router, handlers ...router.Constructor) {
-	log.Debug("Loading OAuth Endpoints")
-
-	// Oauth servers endpoints
-	oAuthHandler := oauth.NewController(p.AuthRepo, p.Notifier)
-	oauthGroup := router.Group("/oauth/servers")
-	{
-		oauthGroup.GET("/", oAuthHandler.Get(), handlers...)
-		oauthGroup.GET("/{name}", oAuthHandler.GetBy(), handlers...)
-
-		if false == p.ReadOnly {
-			oauthGroup.POST("/", oAuthHandler.Post(), handlers...)
-			oauthGroup.PUT("/{name}", oAuthHandler.PutBy(), handlers...)
-			oauthGroup.DELETE("/{name}", oAuthHandler.DeleteBy(), handlers...)
-		}
+		group.POST("/", handler.Post())
+		group.PUT("/{name}", handler.PutBy())
+		group.DELETE("/{name}", handler.DeleteBy())
 	}
 }